Standardise Team.AuthorId type throughout codebase

+++ app/src/main/kotlin/team/finder/api/auth/JwtRequestFilter.kt
@@ -60,7 +60,7 @@ class JwtRequestFilter(val mapper: ObjectMapper, val verifier: JWSVerifier) : On
         }
 
         val idClaim = parsedToken.jwtClaimsSet.getClaim("id")
-        val id = idClaim.toString().toLong()
+        val id = idClaim.toString()
         val subClaim = parsedToken.jwtClaimsSet.subject
 
         val details = CustomUserDetails(id, subClaim, mutableMapOf())
